pyproject.toml
[project]
name = "evo-extraction"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scipy>=1.17.1",
]
